backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
sqlalchemy>=2.0.36
requests>=2.32.0
google-genai>=1.0.0
groq>=0.11.0
pytesseract>=0.3.13
Pillow>=11.0.0
